+++ packages/yourloops/test/integration/pages/product-labelling/product-labelling.common.spec.tsx
@@ -54,50 +54,37 @@ describe('Product labelling page', () => {
     await waitFor(() => {
       expect(router.state.location.pathname).toEqual('/product-labelling')
     })
-
-    expect(screen.getByText(`YourLoops, version ${global.BUILD_CONFIG.VERSION as string}, released on 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-en-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_EN_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Intended Purpose and regulatory information')).toBeInTheDocument()
-    expect(screen.getByText('Legal Manufacturer')).toBeInTheDocument()
+    const udi = screen.getByTestId('udipdf')
+    // doc name should be `YLPZ-RA-LAD-001-fr-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_FR_REV as string}
+    expect(udi).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-en-Rev9.pdf')
+    expect(udi).toHaveAttribute('type', 'application/pdf')
+    expect(udi).toHaveAttribute('height', '100%')
+    expect(udi).toHaveAttribute('width', '100%')
 
     act(() => {
       i18n.changeLanguage(LanguageCodes.Fr)
     })
-    expect(screen.getByText(`YourLoops, version ${global.BUILD_CONFIG.VERSION as string}, publiée le 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-fr-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_FR_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Usage prévu et informations réglementaires')).toBeInTheDocument()
-    expect(screen.getByText('Fabricant')).toBeInTheDocument()
-
-    act(() => {
-      i18n.changeLanguage(LanguageCodes.It)
-    })
-    expect(screen.getByText(`YourLoops, versione ${global.BUILD_CONFIG.VERSION as string}, rilasciato il 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-it-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_IT_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Scopo previsto')).toBeInTheDocument()
-    expect(screen.getByText('Fabricant')).toBeInTheDocument()
-
-    act(() => {
-      i18n.changeLanguage(LanguageCodes.Es)
-    })
-    expect(screen.getByText(`YourLoops, versión ${global.BUILD_CONFIG.VERSION as string}, publicada el 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-es-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_ES_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Finalidad prevista e información regulatoria')).toBeInTheDocument()
-    expect(screen.getByText('Fabricante légal')).toBeInTheDocument()
+    expect(screen.getByTestId('udipdf')).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-fr-Rev9.pdf')
 
     act(() => {
       i18n.changeLanguage(LanguageCodes.Nl)
     })
-    expect(screen.getByText(`YourLoops, versie ${global.BUILD_CONFIG.VERSION as string}, uitgebracht op 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-nl-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_NL_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Beoogd doel')).toBeInTheDocument()
-    expect(screen.getByText('Wettelijk fabrikant')).toBeInTheDocument()
+    expect(screen.getByTestId('udipdf')).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-nl-Rev6.pdf')
+
+    act(() => {
+      i18n.changeLanguage(LanguageCodes.Es)
+    })
+    expect(screen.getByTestId('udipdf')).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-es-Rev6.pdf')
+
+    act(() => {
+      i18n.changeLanguage(LanguageCodes.It)
+    })
+    expect(screen.getByTestId('udipdf')).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-it-Rev6.pdf')
 
     act(() => {
       i18n.changeLanguage(LanguageCodes.De)
     })
-    expect(screen.getByText(`YourLoops, version ${global.BUILD_CONFIG.VERSION as string}, veröffentlicht am 2000-01-01`)).toBeInTheDocument()
-    expect(screen.getByText(`YLPZ-RA-LAD-001-de-Rev${global.BUILD_CONFIG.YLPZ_RA_LAD_001_DE_REV as string}`)).toBeInTheDocument()
-    expect(screen.getByText('Verwendungszweck')).toBeInTheDocument()
-    expect(screen.getByText('Rechtmäßiger Hersteller')).toBeInTheDocument()
+    expect(screen.getByTestId('udipdf')).toHaveAttribute('data', 'fake-urlYLPZ-RA-LAD-001-de-Rev6.pdf')
+
   })
 })